+++ src/lib/command/CommandManager.cpp
@@ -40,8 +40,9 @@ wxBEGIN_EVENT_TABLE(CommandManager, wxEvtHandler)
     EVT_MENU(+CommandId::CloseAll,     CommandManager::onCloseAll)
     EVT_MENU(+CommandId::NewWindow,    CommandManager::onNewWindow)
     EVT_MENU(+CommandId::Quit,         CommandManager::onQuit)
-    EVT_MENU(+CommandId::SessionNew,   CommandManager::onSessionNew)
-    EVT_MENU(+CommandId::SessionClose, CommandManager::onSessionClose)
+    EVT_MENU(+CommandId::SessionNew,    CommandManager::onSessionNew)
+    EVT_MENU(+CommandId::SessionSaveAs, CommandManager::onSessionSaveAs)
+    EVT_MENU(+CommandId::SessionClose,  CommandManager::onSessionClose)
     EVT_MENU_RANGE(wxID_FILE1, wxID_FILE9, CommandManager::onFileHistory)
 
     // Edit
@@ -143,6 +144,7 @@ CommandManager::CommandManager(Context& ctx)
         CommandEntry { .id = +CommandId::SelectAll,        .name="selectAll" },
         CommandEntry { .id = +CommandId::SelectLine,       .name="selectLine" },
         CommandEntry { .id = +CommandId::SessionNew,       .name="sessionNew" },
+        CommandEntry { .id = +CommandId::SessionSaveAs,    .name="sessionSaveAs", .enabled = false },
         CommandEntry { .id = +CommandId::SessionClose,     .name="sessionClose", .enabled = false },
         CommandEntry { .id = +CommandId::ShowExitCode,     .name="showExitCode", .kind = wxITEM_CHECK },
         CommandEntry { .id = +CommandId::Subs,             .name="viewSubs" },
@@ -239,6 +241,10 @@ void CommandManager::onQuit(wxCommandEvent&) {
 
 void CommandManager::onSessionNew(wxCommandEvent&) {
     m_ctx.getDocumentManager().newSession();
+}
+
+void CommandManager::onSessionSaveAs(wxCommandEvent&) {
+    m_ctx.getDocumentManager().saveSessionAs();
 }
 
 void CommandManager::onSessionClose(wxCommandEvent&) {